+++ x/marker/keeper/keeper.go
@@ -192,3 +192,18 @@ func (k Keeper) GetEscrow(ctx sdk.Context, marker types.MarkerAccountI) sdk.Coin
 func (k Keeper) GetAuthority() string {
 	return k.authority
 }
+
+func (k Keeper) IsSendDeny(ctx sdk.Context, markerAddr, senderAddr sdk.AccAddress) bool {
+	store := ctx.KVStore(k.storeKey)
+	return store.Has(types.DenySendKey(markerAddr, senderAddr))
+}
+
+func (k Keeper) AddSendDeny(ctx sdk.Context, markerAddr, senderAddr sdk.AccAddress) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.DenySendKey(markerAddr, senderAddr), []byte{})
+}
+
+func (k Keeper) RemoveSendDeny(ctx sdk.Context, markerAddr, senderAddr sdk.AccAddress) {
+	store := ctx.KVStore(k.storeKey)
+	store.Delete(types.DenySendKey(markerAddr, senderAddr))
+}